+++ nestjs/src/database/repository.ts
@@ -1,4 +1,4 @@
-import { EntityManager } from '../entityManager.service';
+import { EntityManager } from './entityManager.service';
 
 export abstract class Repository {
   protected constructor(readonly entityManager: EntityManager) {}
