+++ nestjs/src/modules/auth/auth.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import {AuthGuard} from "./auth.guard";
+
+@Module({
+    providers: [AuthGuard],
+})
+export class AuthModule {}
