+++ nestjs/src/service/workspaceUser.service.ts
@@ -1,0 +1,12 @@
+import { INestApplication, Injectable } from '@nestjs/common';
+
+@Injectable()
+export class WorkspaceUserService {
+
+    constructor(private workspaceUserRepository: WorkspaceUserService) {
+    }
+
+    createWorkspaceUser() {
+
+    }
+}
